refactor(migrate): split migration assembly
Move citum-migrate final style assembly out of the binary entrypoint
and make the full-template fixup boundary explicit before type-variant
compression.
